Give fork unique name (better suggestion welcome)
