+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="app.olauncher">
+    package="app.olaunchercf">
 
     <uses-permission android:name="android.permission.SET_WALLPAPER" />
     <uses-permission android:name="android.permission.EXPAND_STATUS_BAR" />
@@ -18,7 +18,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".MainActivity"
+            android:name="app.olaunchercf.MainActivity"
             android:configChanges="uiMode"
             android:excludeFromRecents="true"
             android:launchMode="singleTask"
@@ -34,7 +34,7 @@
         </activity>
 
         <activity
-            android:name=".helper.FakeHomeActivity"
+            android:name="app.olaunchercf.helper.FakeHomeActivity"
             android:enabled="false"
             android:launchMode="singleTask"
             android:windowSoftInputMode="stateHidden">
@@ -48,7 +48,7 @@
         </activity>
 
         <receiver
-            android:name=".listener.DeviceAdmin"
+            android:name="app.olaunchercf.listener.DeviceAdmin"
             android:permission="android.permission.BIND_DEVICE_ADMIN">
             <meta-data
                 android:name="android.app.device_admin"
@@ -59,7 +59,7 @@
         </receiver>
 
         <service
-            android:name=".helper.MyAccessibilityService"
+            android:name="app.olaunchercf.helper.MyAccessibilityService"
             android:label="@string/app_name"
             android:permission="android.permission.BIND_ACCESSIBILITY_SERVICE"
             android:process=":serviceProcess">
