+++ app/src/main/java/app/olaunchercf/data/AppModel.kt
@@ -1,4 +1,4 @@
-package app.olauncher.data
+package app.olaunchercf.data
 
 import android.os.UserHandle
 import java.text.CollationKey
