+++ app/src/main/java/app/olaunchercf/data/Prefs.kt
@@ -1,4 +1,4 @@
-package app.olauncher.data
+package app.olaunchercf.data
 
 import android.content.Context
 import android.content.SharedPreferences
