+++ app/src/main/java/app/olaunchercf/helper/MyAccessibilityService.kt
@@ -1,4 +1,4 @@
-package app.olauncher.helper
+package app.olaunchercf.helper
 
 import android.accessibilityservice.AccessibilityService
 import android.content.Intent
@@ -6,8 +6,8 @@ import android.os.Build
 import android.view.accessibility.AccessibilityEvent
 import android.view.accessibility.AccessibilityNodeInfo
 import androidx.annotation.RequiresApi
-import app.olauncher.R
-import app.olauncher.data.Prefs
+import app.olaunchercf.R
+import app.olaunchercf.data.Prefs
 
 class MyAccessibilityService : AccessibilityService() {
