+++ app/src/main/java/app/olaunchercf/listener/OnSwipeTouchListener.kt
@@ -1,4 +1,4 @@
-package app.olauncher.listener
+package app.olaunchercf.listener
 
 import android.content.Context
 import android.view.GestureDetector
@@ -6,20 +6,28 @@ import android.view.GestureDetector.SimpleOnGestureListener
 import android.view.MotionEvent
 import android.view.View
 import android.view.View.OnTouchListener
+import app.olaunchercf.data.Constants
+import java.util.*
+import kotlin.concurrent.schedule
 import kotlin.math.abs
 
-internal open class ViewSwipeTouchListener(c: Context?, v: View) : OnTouchListener {
+/*
+Swipe, double tap and long press touch listener for a view
+Source: https://www.tutorialspoint.com/how-to-handle-swipe-gestures-in-kotlin
+*/
+
+internal open class OnSwipeTouchListener(c: Context?) : OnTouchListener {
+    private var longPressOn = false
+    private var doubleTapOn = false
     private val gestureDetector: GestureDetector
 
     override fun onTouch(view: View, motionEvent: MotionEvent): Boolean {
-        when (motionEvent.action) {
-            MotionEvent.ACTION_DOWN -> view.isPressed = true
-            MotionEvent.ACTION_UP -> view.isPressed = false
-        }
+        if (motionEvent.action == MotionEvent.ACTION_UP)
+            longPressOn = false
         return gestureDetector.onTouchEvent(motionEvent)
     }
 
-    private inner class GestureListener(private val view: View) : SimpleOnGestureListener() {
+    private inner class GestureListener : SimpleOnGestureListener() {
         private val SWIPE_THRESHOLD: Int = 100
         private val SWIPE_VELOCITY_THRESHOLD: Int = 100
 
@@ -28,17 +36,29 @@ internal open class ViewSwipeTouchListener(c: Context?, v: View) : OnTouchListen
         }
 
         override fun onSingleTapUp(e: MotionEvent): Boolean {
-            onClick(view)
+            if (doubleTapOn) {
+                doubleTapOn = false
+                onTripleClick()
+            }
             return super.onSingleTapUp(e)
         }
 
         override fun onDoubleTap(e: MotionEvent): Boolean {
-            onDoubleClick()
+            doubleTapOn = true
+            Timer().schedule(Constants.TRIPLE_TAP_DELAY_MS.toLong()) {
+                if (doubleTapOn) {
+                    doubleTapOn = false
+                    onDoubleClick()
+                }
+            }
             return super.onDoubleTap(e)
         }
 
         override fun onLongPress(e: MotionEvent) {
-            onLongClick(view)
+            longPressOn = true
+            Timer().schedule(Constants.LONG_PRESS_DELAY_MS.toLong()) {
+                if (longPressOn) onLongClick()
+            }
             super.onLongPress(e)
         }
 
@@ -71,11 +91,12 @@ internal open class ViewSwipeTouchListener(c: Context?, v: View) : OnTouchListen
     open fun onSwipeLeft() {}
     open fun onSwipeUp() {}
     open fun onSwipeDown() {}
-    open fun onLongClick(view: View) {}
-    private fun onDoubleClick() {}
-    open fun onClick(view: View) {}
+    open fun onLongClick() {}
+    open fun onDoubleClick() {}
+    open fun onTripleClick() {}
+    private fun onClick() {}
 
     init {
-        gestureDetector = GestureDetector(c, GestureListener(v))
+        gestureDetector = GestureDetector(c, GestureListener())
     }
 }
