+++ app/src/main/java/app/olaunchercf/listener/ViewSwipeTouchListener.kt
@@ -1,4 +1,4 @@
-package app.olauncher.listener
+package app.olaunchercf.listener
 
 import android.content.Context
 import android.view.GestureDetector
@@ -6,28 +6,20 @@ import android.view.GestureDetector.SimpleOnGestureListener
 import android.view.MotionEvent
 import android.view.View
 import android.view.View.OnTouchListener
-import app.olauncher.data.Constants
-import java.util.*
-import kotlin.concurrent.schedule
 import kotlin.math.abs
 
-/*
-Swipe, double tap and long press touch listener for a view
-Source: https://www.tutorialspoint.com/how-to-handle-swipe-gestures-in-kotlin
-*/
-
-internal open class OnSwipeTouchListener(c: Context?) : OnTouchListener {
-    private var longPressOn = false
-    private var doubleTapOn = false
+internal open class ViewSwipeTouchListener(c: Context?, v: View) : OnTouchListener {
     private val gestureDetector: GestureDetector
 
     override fun onTouch(view: View, motionEvent: MotionEvent): Boolean {
-        if (motionEvent.action == MotionEvent.ACTION_UP)
-            longPressOn = false
+        when (motionEvent.action) {
+            MotionEvent.ACTION_DOWN -> view.isPressed = true
+            MotionEvent.ACTION_UP -> view.isPressed = false
+        }
         return gestureDetector.onTouchEvent(motionEvent)
     }
 
-    private inner class GestureListener : SimpleOnGestureListener() {
+    private inner class GestureListener(private val view: View) : SimpleOnGestureListener() {
         private val SWIPE_THRESHOLD: Int = 100
         private val SWIPE_VELOCITY_THRESHOLD: Int = 100
 
@@ -36,29 +28,17 @@ internal open class OnSwipeTouchListener(c: Context?) : OnTouchListener {
         }
 
         override fun onSingleTapUp(e: MotionEvent): Boolean {
-            if (doubleTapOn) {
-                doubleTapOn = false
-                onTripleClick()
-            }
+            onClick(view)
             return super.onSingleTapUp(e)
         }
 
         override fun onDoubleTap(e: MotionEvent): Boolean {
-            doubleTapOn = true
-            Timer().schedule(Constants.TRIPLE_TAP_DELAY_MS.toLong()) {
-                if (doubleTapOn) {
-                    doubleTapOn = false
-                    onDoubleClick()
-                }
-            }
+            onDoubleClick()
             return super.onDoubleTap(e)
         }
 
         override fun onLongPress(e: MotionEvent) {
-            longPressOn = true
-            Timer().schedule(Constants.LONG_PRESS_DELAY_MS.toLong()) {
-                if (longPressOn) onLongClick()
-            }
+            onLongClick(view)
             super.onLongPress(e)
         }
 
@@ -91,12 +71,11 @@ internal open class OnSwipeTouchListener(c: Context?) : OnTouchListener {
     open fun onSwipeLeft() {}
     open fun onSwipeUp() {}
     open fun onSwipeDown() {}
-    open fun onLongClick() {}
-    open fun onDoubleClick() {}
-    open fun onTripleClick() {}
-    private fun onClick() {}
+    open fun onLongClick(view: View) {}
+    private fun onDoubleClick() {}
+    open fun onClick(view: View) {}
 
     init {
-        gestureDetector = GestureDetector(c, GestureListener())
+        gestureDetector = GestureDetector(c, GestureListener(v))
     }
 }
