+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Olauncher</string>
+    <string name="app_name">Olauncher Clutter Free</string>
     <string name="app">App</string>
     <string name="set_as_default_launcher">Set as default launcher</string>
     <string name="app_text_color">Text color</string>
@@ -72,9 +72,4 @@
     <string name="open">Open link</string>
     <string name="copy">Copy</string>
     <string name="close">Close</string>
-    <string name="keeping_it_ad_free">Keeping it Ad-Free</string>
-    <string name="public_roadmap">Olauncher public roadmap</string>
-    <string name="instagram">Instagram</string>
-    <string name="twitter">Twitter</string>
-    <string name="roadmap">Roadmap</string>
 </resources>
